LayoutTests/http/tests/security/xssAuditor/javascript-link-safe-expected.txt

CONSOLE MESSAGE: This is a safe script.
